+++ src/ULinkGame.Server/Actors/ActorDirectoryUnregisterStatus.cs
@@ -1,0 +1,8 @@
+namespace ULinkGame.Server.Actors;
+
+public enum ActorDirectoryUnregisterStatus
+{
+    Unregistered,
+    NotFound,
+    OwnershipMismatch
+}